+++ processor/src/test/resources/input/com/example/prop/ImmutablePropTypeActAsComponentClass.java
@@ -1,0 +1,28 @@
+package com.example.prop;
+
+import arez.annotations.ActAsComponent;
+import javax.annotation.Nullable;
+import react4j.Component;
+import react4j.ReactNode;
+import react4j.annotations.Prop;
+import react4j.annotations.ReactComponent;
+
+@ReactComponent
+abstract class ImmutablePropTypeActAsComponentClass
+  extends Component
+{
+  @ActAsComponent
+  static abstract class MyComponent
+  {
+  }
+
+  @Prop( immutable = true )
+  protected abstract MyComponent getMyProp();
+
+  @Nullable
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+}